Clear remaining MSVC narrowing/discard warnings; build is now warning-free
Make the remaining implicit conversions explicit across the cold/boundary code
so the MSVC build compiles with zero warnings. As before, no types were widened:
each cast is the same truncation the compiler already emitted, so codegen and
runtime behavior are unchanged. Rationale per warning class:

- C4267 size_t->u32/int at API boundaries: GPU buffer byte counts (Buffer.h),
  file write sizes (File.h), hash input size (Hash.h), OpenCL wait-list/program
  counts and workgroup size (clwrap.cpp, KernelCompiler.cpp), trig table sizes
  (TrigBufCache.cpp), and container-index loops (tune.cpp, TuneEntry.cpp,
  state.cpp). These all pass through narrow OpenCL/CRT APIs that already take
  32-bit counts; the values are bounded well within u32.
- C4244 double/float/u64 narrowings: float weight/trig rounding and z-score
  values (TrigBufCache.cpp, tune.cpp), bit/word helpers (shared.h, state.h),
  and the EXP kernel define in Gpu.cpp (the OpenCL side uses a u32 exponent;
  >2^32 support is intentionally not present yet).
- C4018 signed/unsigned: make the (cond ? 1 : 2) loop bounds unsigned (1u/2u)
  to match the u32 loop counter in TrigBufCache.cpp.
- C4834 [[nodiscard]] discarded: explicitly (void)-cast the std::stoi call in
  typeName.h whose result is intentionally ignored (only the parsed length is
  used).

shared.h is included from both C++ and OpenCL, so its casts use C-style
(u32)(...) rather than function-style to stay valid in OpenCL C.

Verified correctness-neutral and warning-free: a clean MSVC Release rebuild
produces 0 warnings/0 errors, and res64 + Z are bit-identical to the recorded
FP64 1K:5:256 / -prp 19999999 baseline at the 2000 and 10000 iter checkpoints.

+++ src/Buffer.h
@@ -38,7 +38,7 @@ private:
   void fill(T value, u32 sizeOrFull = 0) {
     assert(sizeOrFull <= size);
     auto fillSize = sizeOrFull ? sizeOrFull : size;
-    queue->fillBuf(get(), value, fillSize * sizeof(T), tInfo);
+    queue->fillBuf(get(), value, u32(fillSize * sizeof(T)), tInfo);
   }
 
 public:
@@ -62,7 +62,7 @@ public:
 
   void read(T* out, size_t readSize) const {
     assert(readSize && readSize <= size);
-    queue->readSync(get(), readSize * sizeof(T), out, tInfo);
+    queue->readSync(get(), u32(readSize * sizeof(T)), out, tInfo);
   }
 
   // sync read without memory alloc
@@ -80,13 +80,13 @@ public:
     auto readSize = sizeOrFull ? sizeOrFull : size;
     assert(readSize <= size);
     out.resize(readSize);
-    queue->readAsync(get(), readSize * sizeof(T), out.data(), tInfo);
+    queue->readAsync(get(), u32(readSize * sizeof(T)), out.data(), tInfo);
   }
 
   void write(const vector<T>& vect) { queue->write(get(), vect, tInfo); }
 
   void zero(size_t sizeOrFull = 0) {
-    fill(0, sizeOrFull);
+    fill(0, u32(sizeOrFull));
   }
 
   void set(T value) {
@@ -97,7 +97,7 @@ public:
   // device-side copy
   void operator<<(const Buffer<T>& rhs) {
     assert(size == rhs.size);
-    queue->copyBuf(rhs.get(), get(), size * sizeof(T), tInfo);
+    queue->copyBuf(rhs.get(), get(), u32(size * sizeof(T)), tInfo);
     // copyBuf(queue->get(), rhs.get(), get(), size * sizeof(T));
   }
 };
